requirements.txt
torch==2.3.1
torchvision==0.18.1
transformers==4.41.2
datasets==2.19.2
tokenizers==0.19.1
scikit-learn==1.5.0
numpy==1.26.4
matplotlib==3.9.0
tqdm==4.66.4
pyyaml==6.0.1
lm-eval==0.4.3
accelerate==0.30.1
evaluate==0.4.2
hypothesis==6.103.1
pytest==8.2.2
